perf: optimize zhihu sync time
