+++ src/background.js
@@ -1367,115 +1367,7 @@ async function syncToPlatform(platformId, content) {
 
     // ==== 以下是原有的平台特定逻辑（待迁移）====
 
-    if (platformId === 'zhihu') {
-      // 知乎：使用导入文档功能上传 md 文件
-      tab = await chrome.tabs.create({ url: platform.publishUrl, active: false })
-      await addTabToSyncGroup(tab.id, tab.windowId)
-      await waitForTab(tab.id)
-
-      // 等待页面完全加载
-      await new Promise(resolve => setTimeout(resolve, 3000))
-
-      // 在页面中执行：点击导入文档并上传 md 文件
-      await chrome.scripting.executeScript({
-        target: { tabId: tab.id },
-        func: (title, markdown) => {
-          // 等待元素出现的工具函数
-          function waitFor(selector, timeout = 10000) {
-            return new Promise((resolve) => {
-              const start = Date.now()
-              const check = () => {
-                const el = document.querySelector(selector)
-                if (el) resolve(el)
-                else if (Date.now() - start > timeout) resolve(null)
-                else setTimeout(check, 200)
-              }
-              check()
-            })
-          }
-
-          async function uploadMarkdown() {
-            // 先填充标题
-            const titleInput = await waitFor('textarea[placeholder*="标题"]')
-            if (titleInput && title) {
-              titleInput.focus()
-              titleInput.value = title
-              titleInput.dispatchEvent(new Event('input', { bubbles: true }))
-              console.log('[COSE] 知乎标题填充成功')
-            }
-
-            // 第一步：点击工具栏的"导入"按钮，打开子菜单
-            // 注意：按钮文本可能包含零宽字符，使用 includes 匹配
-            const importBtn = Array.from(document.querySelectorAll('button'))
-              .find(el => el.innerText.includes('导入') && !el.innerText.includes('导入文档') && !el.innerText.includes('导入链接'))
-
-            if (importBtn) {
-              importBtn.click()
-              console.log('[COSE] 已点击导入按钮')
-
-              // 等待子菜单出现
-              await new Promise(resolve => setTimeout(resolve, 500))
-
-              // 第二步：点击子菜单中的"导入文档"按钮
-              const importDocBtn = Array.from(document.querySelectorAll('button'))
-                .find(el => el.innerText.includes('导入文档'))
-
-              if (importDocBtn) {
-                importDocBtn.click()
-                console.log('[COSE] 已点击导入文档按钮')
-
-                // 等待上传对话框出现
-                await new Promise(resolve => setTimeout(resolve, 1000))
-
-                // 查找接受 md 文件的输入框
-                const fileInput = document.querySelector('input[type="file"][accept*=".md"]')
-                if (fileInput) {
-                  // 创建 md 文件（使用 text/plain 类型，更通用）
-                  const mdContent = markdown || ''
-                  const fileName = (title || 'article').replace(/[\\/:*?"<>|]/g, '_') + '.md'
-                  const file = new File([mdContent], fileName, { type: 'text/plain' })
-
-                  // 创建 DataTransfer 并设置文件
-                  const dt = new DataTransfer()
-                  dt.items.add(file)
-                  fileInput.files = dt.files
-
-                  // 触发 input 和 change 事件
-                  fileInput.dispatchEvent(new Event('input', { bubbles: true }))
-                  fileInput.dispatchEvent(new Event('change', { bubbles: true }))
-                  console.log('[COSE] 已上传 md 文件:', fileName)
-
-                  // 如果 change 不起作用，尝试拖放方式
-                  await new Promise(resolve => setTimeout(resolve, 500))
-
-                  // 查找上传按钮区域并触发拖放
-                  const dropZone = document.querySelector('[class*="Modal"]') || document.body
-                  const dropEvent = new DragEvent('drop', {
-                    bubbles: true,
-                    cancelable: true,
-                    dataTransfer: dt
-                  })
-                  dropZone.dispatchEvent(dropEvent)
-                  console.log('[COSE] 已触发拖放事件')
-                } else {
-                  console.log('[COSE] 未找到文件输入框')
-                }
-              } else {
-                console.log('[COSE] 未找到导入文档按钮')
-              }
-            } else {
-              console.log('[COSE] 未找到导入按钮')
-            }
-          }
-
-          uploadMarkdown().catch(console.error)
-        },
-        args: [content.title, content.markdown],
-        world: 'MAIN',
-      })
-
-      return { success: true, message: '已打开知乎并导入文档', tabId: tab.id }
-    } else if (platformId === 'infoq') {
+    if (platformId === 'infoq') {
       // InfoQ：需要先调用 API 创建草稿获取 ID，不能直接访问 /draft/write
       try {
         // 调用创建草稿 API
